feat: add `img` inset shadow and hover effect
Inset box shadows on `img` are not natively supported so it's hacky.
`img` is wrapped in a `div` with a `div` sibling. `z-index` is only
compared to sibling elements with parent's being placed behind children.
Ergo, a `div` must be manually overlapped with `img` and have a greater
`z-index` in order for the inset shadow to apply.